requirements.txt
# Build dependencies
setuptools>=65.5.0
wheel

# Core dependencies (updated for Python 3.12 compatibility)
numpy>=1.24.0
pyparsing>=2.4.7
jupyterlab>=3.0.14
plotly>=5.6.0
ipywidgets>=7.6.5
scikit-image>=0.20.0
scikit-learn>=1.3.0
torch
torchvision
torchaudio

# Pip dependencies (relaxed versions for compatibility)
absl-py>=1.0.0
antlr4-python3-runtime>=4.8
cachetools>=5.0.0
charset-normalizer>=2.0.12
cloudpickle>=1.6.0
cycler>=0.11.0
filelock>=3.0.12
fonttools>=4.29.1
geoopt>=0.4.1
glcontext>=2.3.4
glfw>=1.12.0
google-auth>=2.6.0
google-auth-oauthlib>=0.4.6
grpcio>=1.44.0
gym>=0.18.3
gym3>=0.3.3
hydra-core>=1.1.1
idna>=3.3
imageio>=2.9.0
imageio-ffmpeg>=0.3.0
kiwisolver>=1.3.2
markdown>=3.3.6
matplotlib>=3.1.2
moderngl>=5.6.4
oauthlib>=3.2.0
omegaconf>=2.1.1
pandas>=1.4.1
procgen-mirror>=0.10.4
protobuf>=3.19.4
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pyglet>=1.5.15
pytz>=2021.3
pyyaml>=6.0
requests>=2.27.1
requests-oauthlib>=1.3.1
rsa>=4.8
scipy>=1.9.0
seaborn>=0.9.0
tensorboard>=2.8.0
tensorboard-data-server>=0.6.1
tensorboard-plugin-wit>=1.8.1
tqdm>=4.62.3
urllib3>=1.26.8
werkzeug>=2.0.3
hyperbolic>=1.4.0
drawSvg>=1.8.3
